requirements.txt
Flask
Flask-SocketIO
RPi.GPIO 